Calculator that uses RMI

+++ src/Calculator.java
@@ -2,8 +2,8 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface Calculator extends Remote {
-    int addNums(String num1, String num2) throws RemoteException;
-    int subtractNums(String num1, String num2) throws RemoteException;
-    int multipleNums(String num1, String num2) throws RemoteException;
-    int divideNums(String num1, String num2) throws RemoteException;
+    float addNums(String num1, String num2) throws RemoteException;
+    float subtractNums(String num1, String num2) throws RemoteException;
+    float multipleNums(String num1, String num2) throws RemoteException;
+    float divideNums(String num1, String num2) throws RemoteException;
 }
